backend/requirements.txt
Flask
flask-cors
pandas
openpyxl
gspread
python-dotenv
google-genai
